+++ firebase_rules/lib/src/database/model/convert.dart
@@ -3,5 +3,5 @@ import 'package:firebase_rules/database.dart';
 /// Access to [RulesString] methods
 extension RulesStringExtension on String {
   /// Access to [RulesString] methods
-  RulesString get rules => throw UnimplementedError();
+  RulesString rules() => throw UnimplementedError();
 }