+++ firebase_rules/readme/language.dart
@@ -2,10 +2,11 @@ import 'package:firebase_rules/firebase.dart';
 
 void example() {
   /// Dart objects can be converted to rules objects by calling `rules` on them
-  ''.rules.range(0, 1);
+  ''.rules().range(0, 1);
 
-  /// Methods called on `rules` types also take `rules` types as arguments
-  [].rules.concat([].rules);
+  /// Methods called on `rules` types also take `rules` types as arguments.
+  /// Calling `.rules()` on an iterable or map also allows for casting.
+  [].rules<RulesString>().concat([].rules());
 
   /// Global rules functions are available on the `rules` object
   rules.string(true);