+++ firebase_rules_generator/test_project/lib/database/match/input.dart
@@ -5,7 +5,7 @@ final databaseRules = [
   Match(
     r'rules/users/$userId',
     read: (userId) => auth != null && auth?.uid == userId,
-    write: (userId) => userId == 'user1'.rules,
+    write: (userId) => userId == 'user1'.rules(),
     // Only allow creation
     validate: (userId) => !data.exists(),
     indexOn: ['uid', 'email'],
@@ -13,10 +13,10 @@ final databaseRules = [
       Match(
         r'contracts/$contractId',
         read: (contractId) =>
-            root.child('users'.rules).child(userId).child(contractId).val() !=
+            root.child('users'.rules()).child(userId).child(contractId).val() !=
             null,
         write: (contractId) =>
-            root.child('users'.rules).child(userId).child(contractId).val() !=
+            root.child('users'.rules()).child(userId).child(contractId).val() !=
             null,
       ),
     ],