+++ firebase_rules_generator/test_project/lib/firebase/namespace/bytes/input.dart
@@ -5,10 +5,10 @@ import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_inte
 import 'package:firebase_rules_convert/firebase_rules_convert.dart';
 
 bool test(FirestoreResource<BlobTest> resource) {
-  final a = rules.parseBytes(r'\342\202\254'.rules);
-  final b = resource.data.blob.rules.size();
-  final c = resource.data.blob.rules.toBase64();
-  final d = resource.data.blob.rules.toHexString();
+  final a = rules.parseBytes(r'\342\202\254'.rules());
+  final b = resource.data.blob.rules().size();
+  final c = resource.data.blob.rules().toBase64();
+  final d = resource.data.blob.rules().toHexString();
   return true;
 }