+++ firebase_rules_generator/test_project/lib/firebase/namespace/path/input.dart
@@ -3,16 +3,16 @@
 import 'package:firebase_rules/firebase.dart';
 
 bool test(RulesPath other) {
-  final a = rules.path('/path/to/resource'.rules);
+  final a = rules.path('/path/to/resource'.rules());
   final b = other['fieldname'];
   final c = other[0];
   final d = other.bind(
     {
       'foo': 'something',
       'bar': 'another',
-    }.rules,
+    }.rules(),
   );
-  final e = rules.path('/path/to/resource'.rules, database: 'default');
+  final e = rules.path('/path/to/resource'.rules(), database: 'default');
   return true;
 }