+++ firebase_rules_generator/test_project/lib/firebase/namespace/request/input.dart
@@ -25,8 +25,8 @@ bool test1(RulesRequest request) {
   final t = request.resource;
   final u = request.time;
   final v = request.auth?.token.customClaim<RulesString>('identifier') ==
-      'admin'.rules;
-  final w = request.auth?.token.customClaim('identifier') == 'admin'.rules;
+      'admin'.rules();
+  final w = request.auth?.token.customClaim('identifier') == 'admin'.rules();
   return true;
 }