+++ firebase_rules_generator/test_project/lib/firebase/namespace/rules/debug/input.dart
@@ -3,7 +3,8 @@
 import 'package:firebase_rules/firebase.dart';
 
 bool test(FirestoreResource resource) {
-  final a = rules.debug(rules.debug(resource.id) == rules.debug('test'.rules));
+  final a =
+      rules.debug(rules.debug(resource.id) == rules.debug('test'.rules()));
   return true;
 }