+++ firebase_rules_generator/test_project/lib/firebase/namespace/rules/firestore/input.dart
@@ -3,18 +3,19 @@
 import 'package:firebase_rules/firebase.dart';
 
 bool test() {
-  final a = rules.firestore.exists(rules.path(r'/path/to/resource'.rules));
-  final b = rules.firestore.existsAfter(rules.path(r'/path/to/resource'.rules));
-  final c = rules.firestore.get(rules.path(r'/path/to/resource'.rules));
+  final a = rules.firestore.exists(rules.path(r'/path/to/resource'.rules()));
+  final b =
+      rules.firestore.existsAfter(rules.path(r'/path/to/resource'.rules()));
+  final c = rules.firestore.get(rules.path(r'/path/to/resource'.rules()));
   final d = rules.firestore
       .get<TestResource>(
-        rules.path(r'/path/to/resource'.rules, database: 'default'),
+        rules.path(r'/path/to/resource'.rules(), database: 'default'),
       )
       .data
       .asdf;
-  final e = rules.firestore.getAfter(rules.path(r'/path/to/resource'.rules));
+  final e = rules.firestore.getAfter(rules.path(r'/path/to/resource'.rules()));
   final f = rules.firestore
-      .getAfter<TestResource>(rules.path(r'/path/to/resource'.rules))
+      .getAfter<TestResource>(rules.path(r'/path/to/resource'.rules()))
       .data
       .asdf;
   return true;