+++ firebase_rules/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Adds a `customClaim<T>()` method to `RulesToken` to allow access to custom claims
 - Functions are now declared on `Match` statements instead of the annotation
 - `firestore.get<T>()` and `firestore.getAfter<T>()` now properly return a `FirestoreResource<T>`
+- The conversion to rules types is now `.rules()` instead of `.rules` to allow for casting iterables and maps
 
 ## 0.0.1+2
 - Readme fixes