+++ firebase_rules_convert/CHANGELOG.md
@@ -1,3 +1,6 @@
+## NEXT
+- Changed call from `.rules` to `.rules()` to match the new API
+
 ## 0.0.1
 - Initial release